requirements.txt
mysql-connector-python==8.0.33
faker==18.13.0
tabulate==0.9.0
tqdm==4.65.0
python-dotenv==1.0.0
